feat: Add SparklesIcon and SmartLog functionality
Introduces a new SparklesIcon for visual flair and integrates the SmartLog feature. This allows users to log meals and other data by simply describing them in natural language, which is then processed by a language model to extract relevant information and update the application's state.

This also includes initial recipe data to populate the MealsTab and enable recipe browsing.